move services/service.ts to services/core/
