+++ app/services-manager.ts
@@ -1,7 +1,7 @@
 import electron from 'electron';
 import uuid from 'uuid/v4';
 import 'reflect-metadata';
-import { Service } from './services/service';
+import { Service } from './services/core/service';
 import { ObsImporterService } from './services/obs-importer';
 import { ScenesService, SceneItem, SceneItemFolder, Scene, SceneItemNode } from './services/scenes';
 import { ClipboardService } from './services/clipboard';
