+++ app/services/file-manager.ts
@@ -1,4 +1,4 @@
-import { Service } from 'services/service';
+import { Service } from 'services/core/service';
 import path from 'path';
 import fs from 'fs';
