+++ app/services/platforms/niconico.ts
@@ -1,4 +1,4 @@
-import { Service } from 'services/service';
+import { Service } from 'services/core/service';
 import { IPlatformService, IStreamingSetting } from '.';
 import { HostsService } from 'services/hosts';
 import { SettingsService } from 'services/settings';
