+++ app/services/protocol-links.ts
@@ -1,4 +1,4 @@
-import { Service } from 'services/service';
+import { Service } from 'services/core/service';
 import electron from 'electron';
 import { URL, URLSearchParams } from 'url';
 import { Inject } from 'services/core/injector';
