+++ app/util/profanity.ts
@@ -1,4 +1,4 @@
-import { Service } from './../services/service';
+import { Service } from '../services/core/service';
 
 interface IProfanityFilterOptions {
   useDefaultRegex?: boolean;
